Fix Type Casting for partitionOffsets to Use Native int
Resolved an issue where partitionOffsets was incorrectly attempting to cast elements as Int64, leading to a compile-time error due to the absence of the Int64 type. This fix adjusts the casting logic to directly use Dart's native int type, aligning with the actual data type returned by reader.readArray(KafkaType.int64).

+++ lib/src/protocol/offset_api.dart
@@ -107,7 +107,7 @@ class OffsetResponse {
         var partitionId = reader.readInt32();
         var errorCode = reader.readInt16();
         var partitionOffsetsRaw = reader.readArray(KafkaType.int64);
-        var partitionOffsets = partitionOffsetsRaw.map<int>((e) => (e as Int64).toInt()).toList();
+        var partitionOffsets = partitionOffsetsRaw.map<int>((e) => e as int).toList();
         offsets.add(TopicOffsets._(topicName, partitionId, errorCode, partitionOffsets as List<int>)); // ignore: STRONG_MODE_DOWN_CAST_COMPOSITE
         partitionCount--;
       }
